+++ apps/hello/tests.py
@@ -1,9 +1,20 @@
 from django.test import TestCase
+from .models import Mycard
 
-# Create your tests here.
 
+class TestMycardModel(TestCase):
+    Skype_str = "yuriy.torhammer"
 
-class SomeTests(TestCase):
-    def test_math(self):
-        "put docstrings in your tests"
-        assert(2 + 2 == 4)
+    def setUp(self):
+        Mycard.objects.create(fName='Yuriy',
+                              lName='Korostelyov',
+                              Skype=self.Skype_str,
+                              bDate="1983-01-13",
+                              Jabber='ykorostelyov@khavr.com'
+                              )
+
+    def test_person(self):
+        """Testing of My card creatig"""
+        mycard = Mycard.objects.get(id='1')
+        self.assertEqual(mycard.Skype, self.Skype_str)
+        self.assertEqual(mycard.Jabber, 'ykorostelyov@khavr.com')
