+++ fortytwo_test_task/settings/common.py
@@ -41,9 +41,8 @@ INSTALLED_APPS = (
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-
-
-    'Mydata',
+    'south',
+    'apps.hello',
 )
 
 MIDDLEWARE_CLASSES = (
@@ -124,11 +123,10 @@ TEMPLATE_DIRS = (
     # Always use forward slashes, even on Windows.
     # Don't forget to use absolute paths, not relative paths.
     os.path.join(BASE_DIR, 'templates'),
-    os.path.join(BASE_DIR+"/apps/Mydata/templates/")
 )
 
 # FIXTURE_DIRS = (
-#    os.path.join(BASE_DIR+"/apps/Mydata/fixtures/")
+#    os.path.join(BASE_DIR+"/apps/hello/fixtures/")
 # )
 
 # Turn off south during test
