+++ fortytwo_test_task/urls.py
@@ -7,7 +7,7 @@ urlpatterns = patterns(
     '',
     # Examples:
     # url(r'^$', 'fortytwo_test_task.views.home', name='home'),
-    url(r'^$', 'Mydata.views.index', name='home'),
+    url(r'^$', 'hello.views.index', name='home'),
     # url(r'^blog/', include('blog.urls')),
 
     url(r'^admin/', include(admin.site.urls)),
